[ty] Move comparison logic out of builder.rs
